requirements/dev.txt
Flask==1.1.1
Flask-Bcrypt
Flask-DebugToolbar
Flask-Login==0.5.0
Flask-HTTPAuth==3.3.0
Flask-WTF==0.14.3
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
flask-marshmallow==0.11.0
flask-smorest==0.19.2
marshmallow-sqlalchemy==0.22.3
marshmallow==3.5.1
requests
webargs==5.5.3
beautifulsoup4==4.4.1
arrow==0.15.5
sentry-sdk[flask]
cloudinary==1.5.0
pytest-xdist
pytest-cov
pytest
sendgrid
